Add tensorboard video support
This makes it easier to evaluate the progress of an agent. One can now
use the logger to record a video which will be sent to tensorboard. To
keep the interface constant a special Video class has been introduced
which can be used to differentiate between other logging data.
Currently the video is ignored by other loggers as it only really makes
sense to log videos to tensorboard.